# ShopClient/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/clients": {
      "get": {
        "tags": [
          "clients"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/clientGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/clientGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/clientGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "clients"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/clientPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/clientPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/clientPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/clientGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/clientGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/clientGetDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/clients/{id}": {
      "get": {
        "tags": [
          "clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/clientGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/clientGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/clientGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/clientPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/clientPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/clientPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/couriers": {
      "get": {
        "tags": [
          "couriers"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/courierGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/courierGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/courierGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "couriers"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/courierPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/courierPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/courierPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/courierGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/courierGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/courierGetDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/couriers/{id}": {
      "get": {
        "tags": [
          "couriers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/courierGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/courierGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/courierGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "couriers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/courierPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/courierPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/courierPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "couriers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "clientGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fio": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "phone_number": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "clientPostDto": {
        "type": "object",
        "properties": {
          "fio": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "phone_number": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "courierGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fio": {
            "type": "string",
            "nullable": true
          },
          "telephone": {
            "type": "string",
            "nullable": true
          },
          "carId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "courierPostDto": {
        "type": "object",
        "properties": {
          "fio": {
            "type": "string",
            "nullable": true
          },
          "telephone": {
            "type": "string",
            "nullable": true
          },
          "carId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}