# server/Dto/clientGetDto.cs
namespace server.Dto;

public class clientGetDto
{
    public int Id { get; set; }
    public string FIO { get; set; } = string.Empty;
    public string address { get; set; } = string.Empty;
    public string phone_number { get; set; } = string.Empty;

}