# server/Dto/clientPostDto.cs
namespace server.Dto;

public class clientPostDto
{
    public string FIO { get; set; } = string.Empty;
    public string address { get; set; } = string.Empty;
    public string phone_number { get; set; } = string.Empty;
}