# server/Dto/courierGetDto.cs
namespace server.Dto;

public class courierGetDto
{
    public int Id { get; set; }
    public string FIO { get; set; } = string.Empty;
    public string telephone { get; set; } = string.Empty;
    public int carId { get; set; }
}