# server/Dto/courierPostDto.cs
namespace server.Dto
{
    public class courierPostDto
    {
        public string FIO { get; set; } = string.Empty;
        public string telephone { get; set; } = string.Empty;
        public int carId { get; set; }
    }
}